[web] Fix html-elements <BR> element
Instead of rendering as a `<div>`, it renders as a `<br>`, as per the docs.
