+++ packages/html-elements/src/Elements.tsx
@@ -1,5 +1,6 @@
 export * from './elements/Headings';
 export * from './elements/Anchor';
+export * from './elements/Breaks';
 export * from './elements/Layout';
 export * from './elements/Text';
 export * from './elements/Rules';
