+++ packages/html-elements/src/elements/Breaks.web.tsx
@@ -1,0 +1,7 @@
+import createElement from 'react-native-web/dist/exports/createElement';
+
+import { TextProps } from '../primitives/Text';
+
+export function BR(props: TextProps) {
+  return createElement('br', props);
+}
